# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GURU HOST API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
  country: string;
  currency: string;
  herokuApiKey: string;
  herokuTeam?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  country: string;
  currency: string;
  gruCredits: number;
  freeDeploymentUsed: boolean;
  herokuApiType: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type DeploymentStatus =
  (typeof DeploymentStatus)[keyof typeof DeploymentStatus];

export const DeploymentStatus = {
  building: "building",
  running: "running",
  paused: "paused",
  crashed: "crashed",
  deleted: "deleted",
} as const;

export interface Bot {
  id: number;
  name: string;
  description: string;
  repoUrl: string;
  imageUrl?: string;
  features?: string[];
  isActive: boolean;
  createdAt: string;
}

export interface Deployment {
  id: string;
  userId: number;
  botId: number;
  appName: string;
  repoUrl: string;
  sessionId: string;
  status: DeploymentStatus;
  herokuAppId?: string;
  createdAt: string;
  updatedAt: string;
  bot?: Bot;
}

export interface EnvVar {
  key: string;
  value: string;
}

export interface CreateDeploymentRequest {
  botId: number;
  repoUrl: string;
  sessionId: string;
  envVars?: EnvVar[];
}

export interface LogsResponse {
  logs: string;
  lines: string[];
}

export interface CreateBotRequest {
  name: string;
  description: string;
  repoUrl: string;
  imageUrl?: string;
  features?: string[];
}

export interface SubmitPaymentRequest {
  screenshotBase64: string;
  screenshotMimeType: string;
  amount: number;
  phoneNumber?: string;
  notes?: string;
}

export type PaymentRequestStatus =
  (typeof PaymentRequestStatus)[keyof typeof PaymentRequestStatus];

export const PaymentRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface PaymentRequest {
  id: number;
  userId: number;
  amount: number;
  status: PaymentRequestStatus;
  screenshotUrl: string;
  phoneNumber?: string;
  notes?: string;
  createdAt: string;
  updatedAt: string;
  user?: User;
}

export interface FundUserRequest {
  amount: number;
  note?: string;
}

export interface PlatformStats {
  totalUsers: number;
  totalDeployments: number;
  activeDeployments: number;
  pendingPayments: number;
  totalRevenue: number;
  botsInCatalog: number;
}

export type GetDeploymentLogsParams = {
  lines?: number;
};

export type SetDeploymentEnvBody = {
  vars: EnvVar[];
};