# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GURU HOST API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input or email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /deployments:
    get:
      operationId: getDeployments
      tags: [deployments]
      summary: Get all deployments for current user
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Create a new bot deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeploymentRequest"
      responses:
        "200":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "402":
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Admin login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users/{id}/fund:
    post:
      operationId: adminFundUser
      tags: [admin]
      summary: Admin - fund a user's account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundUserRequest"
      responses:
        "200":
          description: User funded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/payments/{id}/approve:
    post:
      operationId: adminApprovePayment
      tags: [admin]
      summary: Admin - approve a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Payment approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
  /admin/payments/{id}/reject:
    post:
      operationId: adminRejectPayment
      tags: [admin]
      summary: Admin - reject a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Payment rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
...
  /admin/bots:
    post:
      operationId: adminCreateBot
      tags: [admin]
      summary: Admin - add a new bot to catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBotRequest"
      responses:
        "200":
          description: Bot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /admin/bots/{id}:
    delete:
      operationId: adminDeleteBot
      tags: [admin]
      summary: Admin - delete a bot from catalog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Bot deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        country:
          type: string
        currency:
          type: string
        herokuApiKey:
          type: string
        herokuTeam:
          type: string
      required: [email, password, name, country, currency, herokuApiKey]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AdminLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        name:
          type: string
        country:
          type: string
        currency:
          type: string
        gruCredits:
          type: number
        freeDeploymentUsed:
          type: boolean
        herokuApiType:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required: [id, email, name, country, currency, gruCredits, freeDeploymentUsed, herokuApiType, isAdmin, createdAt]
    Deployment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: number
        botId:
          type: number
        appName:
          type: string
        repoUrl:
          type: string
        sessionId:
          type: string
        status:
          type: string
          enum: [building, running, paused, crashed, deleted]
        herokuAppId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        bot:
          $ref: "#/components/schemas/Bot"
      required: [id, userId, botId, appName, repoUrl, sessionId, status, createdAt, updatedAt]
    CreateDeploymentRequest:
      type: object
      properties:
        botId:
          type: number
        repoUrl:
          type: string
        sessionId:
          type: string
        envVars:
          type: array
          items:
            $ref: "#/components/schemas/EnvVar"
      required: [botId, repoUrl, sessionId]
    EnvVar:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required: [key, value]
...
    Bot:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        repoUrl:
          type: string
        imageUrl:
          type: string
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, description, repoUrl, isActive, createdAt]
    CreateBotRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        repoUrl:
          type: string
        imageUrl:
          type: string
        features:
          type: array
          items:
            type: string
      required: [name, description, repoUrl]
...
    PaymentRequest:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        screenshotUrl:
          type: string
        phoneNumber:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [id, userId, amount, status, screenshotUrl, createdAt, updatedAt]
    FundUserRequest:
      type: object
      properties:
        amount:
          type: number
        note:
          type: string
      required: [amount]
...