# lib/api-zod/src/generated/types/adminLoginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GURU HOST API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginRequest {
  username: string;
  password: string;
}