# lib/api-zod/src/generated/types/bot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GURU HOST API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Bot {
  id: number;
  name: string;
  description: string;
  repoUrl: string;
  imageUrl?: string;
  features?: string[];
  isActive: boolean;
  createdAt: string;
}