# lib/api-zod/src/generated/types/createDeploymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GURU HOST API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EnvVar } from "./envVar";

export interface CreateDeploymentRequest {
  botId: number;
  repoUrl: string;
  sessionId: string;
  envVars?: EnvVar[];
}