# lib/api-zod/src/generated/types/paymentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GURU HOST API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentRequestStatus } from "./paymentRequestStatus";
import type { User } from "./user";

export interface PaymentRequest {
  id: number;
  userId: number;
  amount: number;
  status: PaymentRequestStatus;
  screenshotUrl: string;
  phoneNumber?: string;
  notes?: string;
  createdAt: string;
  updatedAt: string;
  user?: User;
}