# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GURU HOST API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
  country: string;
  currency: string;
  herokuApiKey: string;
  herokuTeam?: string;
}