# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GURU HOST API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  name: string;
  country: string;
  currency: string;
  gruCredits: number;
  freeDeploymentUsed: boolean;
  herokuApiType: string;
  isAdmin: boolean;
  createdAt: string;
}